+++ python_tamer/subroutines.py
@@ -3,12 +3,12 @@ import numpy as np
 def assert_data_shape_24(data,reverse=False) :
     """Simple function to check if first dimension is 24 hours and, if not, reshapes accordingly
     """
-
+    datashape = np.shape(data)
     # TO DO: Could be a big job, but this F ordering is weird and I should reconsider
-    if np.shape(data)[0] != 24 and not reverse: # Checks that first dimension is length 24 (hours in a day) and reshapes if not
-        new_shape = (24, np.shape(data)[0]//24) + np.shape(data)[1:]
-    elif np.shape(data)[0] == 24 and reverse :
-        new_shape = [24 * np.shape(data)[1]] + list(np.shape(data)[2:])
+    if datashape[0] != 24 and not reverse: # Checks that first dimension is length 24 (hours in a day) and reshapes if not
+        new_shape = (24, datashape[0]//24) + datashape[1:]
+    elif reverse :
+        new_shape = [datashape[0] * datashape[1]] + list(datashape[2:])
     data = np.reshape(data,new_shape,order='F')
     return data
 
@@ -75,6 +75,7 @@ def find_nearest(array, value):
 
 
 def convert_swiss_time_to_UTC(input_table,name) :
+    # TO DO: Need a replacement for this, but argument the responsibility of the user?
     def convert_swiss_time_to_UTC_iter(time_in,Date) :
         if Date.month > 3 and Date.month < 11 :
             time_out = time_in 
@@ -85,4 +86,150 @@ def convert_swiss_time_to_UTC(input_table,name) :
         return time_out
     new_time = input_table.apply(lambda x: convert_swiss_time_to_UTC_iter(x[name],x["Date"]),axis='columns')
     return new_time
+
+
+def hist_mean(counts,bin_centers) :
+    """hist_mean calculates the mean of a histogram using numpy functions
+
+    This function is designed to calculate a histogram mean as efficiently as possible
+
+    Args:
+        counts (array): The quantity of numbers within each histogram bin
+        bin_centers (array): The central value of each histogram bin. Note that
+            this can be calculated as bin_edges[:-1]+0.5*np.diff(bin_edges) but
+            we removed this calculation to optimise this function further.
+
+    Returns:
+        float: the mean value of the histogram
+    """
+
+    mean = np.dot(counts, bin_centers) / np.sum(counts)
+    return mean
+
+
+def hist_var(counts,bin_centers) :
+    """hist_var calculates the variance of a histogram
+
+    This function calculates the variance of a histogram as E[X^2] - E[X]^2 using the
+    hist_mean function to efficiently calculate expectation values.
+
+    Args:
+        counts (array): The quantity of numbers within each histogram bin
+        bin_centers (array): The central value of each histogram bin. Note that
+            this can be calculated as bin_edges[:-1]+0.5*np.diff(bin_edges) but
+            we removed this calculation to optimise this function further.
+
+    Returns:
+        float: the variance of the histogram
+    """
+
+    # Not really essential seeing as this would break the units
+    var = hist_mean(counts,bin_centers**2) - hist_mean(counts,bin_centers)**2
+    return var
+
+
+def hist_stdev(counts,bin_centers) :
+    """hist_stdev calculates the standard deviation of a histogram
+
+    This function calculates the variance of a histogram as E[X^2] - E[X]^2 using the
+    hist_mean function to efficiently calculate expectation values. It then returns the
+    square root of the variance.
+
+    Args:
+        counts (array): The quantity of numbers within each histogram bin
+        bin_centers (array): The central value of each histogram bin. Note that
+            this can be calculated as bin_edges[:-1]+0.5*np.diff(bin_edges) but
+            we removed this calculation to optimise this function further.
+
+    Returns:
+        float: the standard deviation of the histogram
+    """
+
+    var = hist_mean(counts,bin_centers**2) - hist_mean(counts,bin_centers)**2
+    return var**0.5
+
+
+def hist_percentile(counts,bin_centers,prct) :
+    """hist_percentile calculates percentiles of histogram data
+
+    This function takes discretised data, typical for histograms, and calculates
+    the user-specified percentile quantity. 
+
+    Args:
+        counts (array): The quantity of numbers within each histogram bin
+        bin_centers (array): The central value of each histogram bin. Note that
+            this can be calculated as bin_edges[:-1]+0.5*np.diff(bin_edges) but
+            we removed this calculation to optimise this function further.
+        prct (float): A fraction betwee 0.0 and 1.0 for the desired percentile.
+
+    Returns:
+        float: The desired percentile value. In cases where the quantity falls 
+            between two bins, their respective central values are averaged.
+    """
+
+    n = np.sum(counts)
+    cumcounts = np.cumsum(counts)
+    # TO DO: Possibly unnecessary, but could probably improve efficiency of
+    # this if statement (e.g. if i==j no need to take average)
+    if prct == 0 :
+        # special case: searching for min
+        j = np.searchsorted(cumcounts,n*prct,side='right')
+        percentile = bin_centers[j]
+    elif prct == 1 : 
+        # special case: searching for max
+        i = np.searchsorted(cumcounts,n*prct)
+        percentile = bin_centers[i]
+    else :
+        i = np.searchsorted(cumcounts,n*prct) 
+        j = np.searchsorted(cumcounts,n*prct,side='right') 
+        percentile = (bin_centers[i] + bin_centers[j])/2
+
+    return percentile   
+
+
+def hist_min(counts,bin_centers) :
+    """hist_min calculates the minimum value of a histogram
+
+    This function finds the minimum value of a histogram.
+    It is built on the some basic functionality as hist_percentile.
+
+    Args:
+        counts (array): The quantity of numbers within each histogram bin
+        bin_centers (array): The central value of each histogram bin. Note that
+            this can be calculated as bin_edges[:-1]+0.5*np.diff(bin_edges) but
+            we removed this calculation to optimise this function further.
+
+    Returns:
+        float: The minimum value of the histogram
+    """
+
+    cumcounts = np.cumsum(counts)
+    j = np.searchsorted(cumcounts,0,side='right')
+    min = bin_centers[j]
+
+    return min
+
+
+def hist_max(counts,bin_centers) :
+    """hist_max calculates the maximum value of a histogram
+
+    This function finds the maximum value of a histogram.
+    It is built on the some basic functionality as hist_percentile.
+
+    Args:
+        counts (array): The quantity of numbers within each histogram bin
+        bin_centers (array): The central value of each histogram bin. Note that
+            this can be calculated as bin_edges[:-1]+0.5*np.diff(bin_edges) but
+            we removed this calculation to optimise this function further.
+
+    Returns:
+        float: The maximum value of the histogram
+    """
+
+    n = np.sum(counts)
+    cumcounts = np.cumsum(counts)
+    i = np.searchsorted(cumcounts,n)
+    max = bin_centers[i]
+
+    return max