+++ test/test_dummy_library.py
@@ -1,8 +1,12 @@
-from python_tamer.library import dummy_func
+import pytest
+
+from python_tamer.library import *
+
+def test_data_load():
+
+    test = ExposureMap(date_selection=pd.date_range(start="2018-01-01",end="2018-01-01"),
+    units = "UVI",
+    statistic = "max"
+    ).collect_data()
 
 
-def test_dummy_func():
-    """Test the correctness of dummy_func.
-    """
-
-    assert dummy_func(3.2) == 3