+++ HISTORY.rst
@@ -2,7 +2,13 @@
 History
 =======
 
-0.1.0 (2021-03-02)
+0.2 (2021-03-04)
+
+* Finishing touches for alpha release. Still no documentation and limited functionality, but significant extentions to ExposureMap functionality compared to 0.1.
+* Added histogram descriptor calculator functions to subroutines.
+
+
+0.1 (2021-03-02)
 ------------------
 
 * Alpha release on github only, no documentation and limited functionality